server/requirements.txt
asgiref==3.7.2
Django==5.0.2
paho-mqtt==2.0.0
sqlparse==0.4.4
